projects/stolen-goat-prospect-agent/requirements.txt
requests>=2.31
anthropic>=0.40
python-dotenv>=1.0
pyyaml>=6.0
rapidfuzz>=3.0
